+++ Android/shared/src/main/java/com/redbadger/crux_core/shared_types/PlatformMsg.java
@@ -1,4 +1,4 @@
-package com.redbadger.rmm.shared_types;
+package com.redbadger.crux_core.shared_types;
 
 
 public abstract class PlatformMsg {
