+++ crux_core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rmm"
+name = "crux_core"
 version = "0.1.0"
 edition = "2021"
