+++ crux_core/src/platform.rs
