+++ crux_core/src/typegen.rs
