+++ crux_core/typegen_extensions/typescript/serde/binarySerializer.ts
