+++ crux_core/typegen_extensions/typescript/tsconfig.json
