+++ shared/uniffi.toml
@@ -1,5 +1,5 @@
 [bindings.kotlin]
-package_name = "com.redbadger.rmm.shared"
+package_name = "com.redbadger.crux_core.shared"
 cdylib_name = "shared"
 
 [bindings.swift]
